feat(media): Phase 1 — MediaAsset, MediaRendition, EntityMedia models
New media app with three models, 27 DB-level constraints, read-only
admin, and 38 tests (36 constraint + 2 model validation). Schema only —
no behavior, no APIs.

MediaAsset: uploaded media metadata (kind, mime_type, dimensions,
status, uploaded_by with on_delete=PROTECT).

MediaRendition: one stored file per asset (original/thumb/display),
identified by uuid. No storage_key field — paths derived at runtime
by build_storage_key() in Phase 3.

EntityMedia: claim-resolved catalog attachment via GenericFK. Partial
unique constraints enforce one primary per entity per category.
clean() validates that the target content type is MediaSupported.

MediaSupported marker mixin in core/models.py, applied to MachineModel.
EntityMedia.clean() uses issubclass() to reject unsupported entity types.

Also widens Claim.object_id to PositiveBigIntegerField to match
BigAutoField PKs. Adds Media.md architecture plan.

+++ backend/apps/provenance/migrations/0001_initial.py
@@ -211,7 +211,7 @@ class Migration(migrations.Migration):
                         verbose_name="ID",
                     ),
                 ),
-                ("object_id", models.PositiveIntegerField()),
+                ("object_id", models.PositiveBigIntegerField()),
                 ("field_name", models.CharField(max_length=255)),
                 (
                     "claim_key",
